# optimize `memorynew` intrinsic for constant length `Memory`