… Compiler/test/irpasses.jl
… doc/src/manual/performance-tips.md
… src/common_symbols2.inc
… src/genericmemory.c
… src/julia.h
… src/llvm-alloc-helpers.cpp
… src/llvm-alloc-opt.cpp
… src/pipeline.cpp
… src/rtutils.c
… src/staticdata.c
… test/arrayops.jl
… test/core.jl