@gbaraldi 2024-10-04
I'd print everyone involved here with the way I showed you yesterday `thing->print(dbgs())`, So print the func and the arguments