@oscardssmith 2024-10-04
This now works! For simple examples like `Memory{Int}(undef, 3)` it's about 7ns compared to 11 previously. Also this version is almost able to be analyzed by our escape analysis to delete the Memory when it doesn't (we just need to teach the llvm-alloc-helpers about `gc_loaded`, and for non `@inbounds` accesses we need to teach LLVM that the boundschecks are dead).