@giordano 2024-10-04
[`FixedSizeArrays.jl`](https://github.com/JuliaArrays/FixedSizeArrays.jl) is going to be much nicer 🙂