@giordano 2024-10-11
Can you please add an llvm pass test for https://github.com/JuliaLang/julia/pull/56030#issue-2570935140 (removing all memory for a simple case where the Memory object doesn't escape)?