@oscardssmith 2024-10-11
Do you want an actual LLVM pass, or can I just write a test for 0 allocations?