@giordano 2024-10-11
I think an llvm test would be more robust, but probably a simple zero-allocation test would do the job as well.