@oscardssmith 2024-10-14
This is now on top of https://github.com/JuliaLang/julia/pull/55995 (to figure out why we weren't optimizing correctly), but other than that, I think this is good to go!