@giordano 2024-10-15
Maybe a test of no allocations in simple cases as discussed above? 🙂