@MasonProtter 2024-10-18
Is there a way to make `memorynew` able to take a pointer so it can be used with `unsafe_wrap`? Or would that just be a separate project without much overlap with this one? 

I ask because it'd be very nice if `unsafe_wrap` could be non-allocating for non-escaping `Memory`.