@oscardssmith 2024-10-18
should be doable, but mostly unrelated. it would also probably be restricted to be not a bitsunion memory, or atomic memory, but I think it belongs in a followup pr