@MasonProtter 2024-10-24
Here's something that segfaults currently on this PR:
```julia-repl
julia> function f()
           m = Memory{Float64}(undef, 32)
           m .= rand.()
           s = 0.0
           for x ∈ m
               s += x
           end
           s
       end
f (generic function with 2 methods)

julia> code_llvm(f, ())
; Function Signature: f()
;  @ REPL[4]:1 within `f`
define double @julia_f_3154() #0 {
L100:
  unreachable
}
```
and
```julia-repl
julia> f()

[90408] signal 11 (128): Segmentation fault
in expression starting at REPL[6]:1
unknown function (ip: 0x7f1a75762e1f) at (unknown file)
unknown function (ip: 0x7ffe8d30a5cf) at (unknown file)
Allocations: 9892774 (Pool: 9892496; Big: 278); GC: 20
zsh: segmentation fault (core dumped)  ./julia

```