@gbaraldi 2024-10-25
Given that this doesn't build with assertions on i'm not surprised