@KristofferC 2024-10-30
Add that as a test case?