@oscardssmith 2024-10-30
A minimized version of the test is added.